+++ odoo/local-src/db2_import/models/db2_importer.py
@@ -372,7 +372,9 @@ class DB2MapperPurchaseOrder(object):
             xmlid = '__import__.purchase_order_line_%s_%s_%s_%s' % (
                 row['ecfsui'], int(row['ecffou']),
                 int(row['ecfsuc']), int(line['dcfnli']))
-            po_lines |= create_or_update(POLine, xmlid, values)
+            line_rec = create_or_update(POLine, xmlid, values)
+            line['odoo_id'] = line_rec.id
+            po_lines |= line_rec
             received_lines.append(line['dcfquc'] <= line['dcfqul'])
             not_received_lines.append(line['dcfqul'] == 0)
         is_received = all(received_lines)
@@ -389,16 +391,23 @@ class DB2MapperPurchaseOrder(object):
             'state': state,
         })
         if is_received:
-            # force received qty in database to avoid to have
+            # force invoiced and received qty in database to avoid to have
             # to create pickings, this needs to be done after state write
             # or it would be recomputed
             query = (
                 "UPDATE purchase_order_line"
-                " SET qty_received = product_qty,"
-                "     qty_invoiced = product_qty"
+                " SET qty_invoiced = product_qty"
                 " WHERE id in ( %s )"
             ) % ','.join(['%s'] * len(po_lines))
             cr.execute(query, po_lines.ids)
+            query = (
+                "UPDATE purchase_order_line"
+                " SET qty_received = %s"
+                " WHERE id = %s"
+            )
+            for line in lines:
+                cr.execute(query, (line['dcfqul'], line['odoo_id']))
+
         elif rec.importer_id.mode == 'final_update':
             new._create_picking()
             if is_partially_received: